refactor(tag/detail): divide to components
